requirements.txt
langchain==0.0.184
PyPDF2==3.0.1
python-dotenv==1.0.0
streamlit==1.18.1
openai==0.27.6
faiss-cpu==1.7.4
altair<5
bardapi~=0.1.18
PyYAML~=6.0
pandas~=2.0.3
plotly~=5.15.0